norminette and start trim

+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: frgonzal <frgonzal@student.42malaga.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/09/22 19:51:07 by frgonzal          #+#    #+#             */
-/*   Updated: 2026/09/22 19:56:16 by frgonzal         ###   ########.fr       */
+/*   Updated: 2026/09/24 20:47:04 by frgonzal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	 strncmp(const char *s1, const char *s2, size_t n)
+int	strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (n == 0)
 		return (0);
@@ -23,4 +23,3 @@ int	 strncmp(const char *s1, const char *s2, size_t n)
 		i++;
 	return (s1[i] - s2[i]);
 }
-
